Create JS_Proof_Project_code.js
This is the program code of the JS Proof  Final Assessment Proof which will create a variable that can hold up a number of products NFTs in it's wallet and three user-defined functions. First Function will create an object and push it in the NFTs array, Second function  will print the metadata of the entire NFTs and the Third function to print the number of NFTs in the wallet.